[IMP] sale_commission: Remove required attribute in settlement field. Added to view
When updating a table, modules that don't have the `required=True` setting of a field in their scope issue an `ALTER TABLE` at least twice per module. This blocks the entire table.

This module was doing that, and since the `res.partner` model is so ubiquitous, this was blocking concurrent updates.

By moving the required attribute to the views, the UX is the same, but concurrent updates work fine again.

@Tecnativa TT27096